indian-farmer-assistance-app/krishi_rag/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.12
langchain-community==0.2.10
faiss-cpu
sentence-transformers
pydantic==2.7.4
docx2txt
python-dotenv

# Audio processing
openai-whisper==20231117
sounddevice==0.4.6
scipy==1.11.4
pyttsx3==2.90
